Convert the Image component to TypeScript
