+++ common/model/tasl.ts
@@ -1,11 +1,11 @@
 import { LicenseType } from './license';
 
 export type Tasl = {
-  title: string | null;
-  author: string | null;
-  sourceName: string | null;
-  sourceLink: string | null;
-  license: LicenseType | null;
-  copyrightHolder: string | null;
-  copyrightLink: string | null;
+  title?: string;
+  author?: string;
+  sourceName?: string;
+  sourceLink?: string;
+  license?: LicenseType;
+  copyrightHolder?: string;
+  copyrightLink?: string;
 };
