+++ content/webapp/components/BookPromo/BookPromo.tsx
@@ -81,7 +81,6 @@ const BookPromo: FunctionComponent<Props> = ({
                 height={image.height || 0}
                 alt=""
                 sizesQueries="(min-width: 1420px) 386px, (min-width: 960px) calc(28.64vw - 15px), (min-width: 600px) calc(50vw - 54px), calc(100vw - 36px)"
-                tasl={null}
               />
             </BookPromoImage>
           )}
